# perf: Add metric for time spent casting in native scan