--- a/native/core/src/execution/operators/scan.rs
+++ b/native/core/src/execution/operators/scan.rs
@@ -39,7 +39,9 @@ use arrow_data::ffi::FFI_ArrowArray;
 use arrow_data::ArrayData;
 use arrow_schema::ffi::FFI_ArrowSchema;
 use arrow_schema::{DataType, Field, Schema, SchemaRef};
-use datafusion::physical_plan::metrics::{BaselineMetrics, ExecutionPlanMetricsSet, MetricsSet};
+use datafusion::physical_plan::metrics::{
+    BaselineMetrics, ExecutionPlanMetricsSet, MetricBuilder, MetricsSet, Time,
+};
 use datafusion::{
     execution::TaskContext,
     physical_expr::*,
@@ -345,16 +347,20 @@ struct ScanStream<'a> {
     baseline_metrics: BaselineMetrics,
     /// Cast options
     cast_options: CastOptions<'a>,
+    /// Timer for cast operations
+    cast_time: Time,
 }
 
 impl<'a> ScanStream<'a> {
     pub fn new(scan: ScanExec, schema: SchemaRef, partition: usize) -> Self {
         let baseline_metrics = BaselineMetrics::new(&scan.metrics, partition);
+        let cast_time = MetricBuilder::new(&scan.metrics).subset_time("cast_time", partition);
         Self {
             scan,
             schema,
             baseline_metrics,
             cast_options: CastOptions::default(),
+            cast_time,
         }
     }
 
@@ -375,7 +381,10 @@ impl<'a> ScanStream<'a> {
             .zip(schema_fields.iter())
             .map(|(column, f)| {
                 if column.data_type() != f.data_type() {
-                    cast_with_options(column, f.data_type(), &self.cast_options)
+                    let mut timer = self.cast_time.timer();
+                    let cast_array = cast_with_options(column, f.data_type(), &self.cast_options);
+                    timer.stop();
+                    cast_array
                 } else {
                     Ok(Arc::clone(column))
                 }