--- a/spark/src/main/scala/org/apache/spark/sql/comet/CometScanExec.scala
+++ b/spark/src/main/scala/org/apache/spark/sql/comet/CometScanExec.scala
@@ -198,12 +198,12 @@ case class CometScanExec(
     // Tracking scan time has overhead, we can't afford to do it for each row, and can only do
     // it for each batch.
     if (supportsColumnar) {
-      Seq(
+      Map(
         "scanTime" -> SQLMetrics.createNanoTimingMetric(
           sparkContext,
           "scan time")) ++ CometMetricNode.scanMetrics(sparkContext)
     } else {
-      Seq.empty
+      Map.empty
     }
   } ++ {
     relation.fileFormat match {